app/requirements.txt
aio-pika==6.8.0
aiormq==3.3.1
asgiref==3.4.1
blis==0.7.4
catalogue==1.0.0
certifi==2021.5.30
charset-normalizer==2.0.4
click==8.0.1
cymem==2.0.5
dataclasses==0.8
fastapi==0.68.0
h11==0.12.0
idna==3.2
importlib-metadata==4.6.4
joblib==1.0.1
multidict==5.1.0
murmurhash==1.0.5
nltk==3.6.2
numpy==1.19.5
pamqp==2.3.0
pandas==1.1.5
pkg_resources==0.0.0
plac==1.1.3
preshed==3.0.5
psycopg2-binary==2.9.1
pydantic==1.8.2
python-dateutil==2.8.2
python-dotenv==0.19.0
python-multipart==0.0.5
pytz==2021.1
regex==2021.8.21
requests==2.26.0
six==1.16.0
spacy==2.3.7
srsly==1.0.5
starlette==0.14.2
thinc==7.4.5
tika==1.24
tqdm==4.62.2
typing-extensions==3.10.0.0
urllib3==1.26.6
uvicorn==0.15.0
wasabi==0.8.2
yarl==1.6.3
zipp==3.5.0
